organize resource node folders
