+++ src/commands/fluxCheck.ts
@@ -1,12 +1,12 @@
 import safesh from 'shell-escape-tag';
 
 import { shell } from 'cli/shell/exec';
-import { ClusterContextNode } from 'ui/treeviews/nodes/clusterContextNode';
+import { ClusterNode } from 'ui/treeviews/nodes/cluster/clusterNode';
 
 /**
  * Runs `flux check` command for selected cluster in the output view.
  * @param clusterNode target cluster node (from tree node context menu)
  */
-export async function fluxCheck(clusterNode: ClusterContextNode) {
+export async function fluxCheck(clusterNode: ClusterNode) {
 	shell.execWithOutput(safesh`flux check --context ${clusterNode.contextName}`);
 }
