+++ src/commands/fluxReconcileWorkload.ts
@@ -3,8 +3,8 @@ import { window } from 'vscode';
 import { fluxTools } from 'cli/flux/fluxTools';
 import { FluxWorkload } from 'types/fluxCliTypes';
 import { Kind } from 'types/kubernetes/kubernetesTypes';
-import { HelmReleaseNode } from 'ui/treeviews/nodes/helmReleaseNode';
-import { KustomizationNode } from 'ui/treeviews/nodes/kustomizationNode';
+import { HelmReleaseNode } from 'ui/treeviews/nodes/workload/helmReleaseNode';
+import { KustomizationNode } from 'ui/treeviews/nodes/workload/kustomizationNode';
 import { refreshSourcesTreeView, refreshWorkloadsTreeView } from 'ui/treeviews/treeViews';
 
 /**
