+++ src/commands/setCurrentKubernetesContext.ts
@@ -1,11 +1,11 @@
 import { setCurrentContext } from 'cli/kubernetes/kubernetesConfig';
-import { ClusterContextNode } from 'ui/treeviews/nodes/clusterContextNode';
+import { ClusterNode } from 'ui/treeviews/nodes/cluster/clusterNode';
 import { refreshAllTreeViews } from 'ui/treeviews/treeViews';
 
 /**
  * Sets Kubernetes context and refreshes tree views if needed.
  */
-export async function setCurrentKubernetesContext(clusterContext: ClusterContextNode): Promise<void> {
+export async function setCurrentKubernetesContext(clusterContext: ClusterNode): Promise<void> {
 	const setContextResult = await setCurrentContext(clusterContext.contextName);
 	if (setContextResult?.isChanged) {
 		refreshAllTreeViews();
