+++ src/ui/treeviews/nodes/source/helmRepositoryNode.ts
@@ -1,6 +1,6 @@
 import { HelmRepository } from 'types/flux/helmRepository';
 import { Kind } from 'types/kubernetes/kubernetesTypes';
-import { NodeContext } from '../../../types/nodeContext';
+import { NodeContext } from '../../../../types/nodeContext';
 import { SourceNode } from './sourceNode';
 
 /**
