+++ src/ui/treeviews/nodes/workload/helmReleaseNode.ts
@@ -1,6 +1,6 @@
 import { HelmRelease } from 'types/flux/helmRelease';
 import { Kind } from 'types/kubernetes/kubernetesTypes';
-import { NodeContext } from '../../../types/nodeContext';
+import { NodeContext } from '../../../../types/nodeContext';
 import { WorkloadNode } from './workloadNode';
 
 /**
